bp: Move bp only state to bp
