+++ BUILD.gn
@@ -77,6 +77,7 @@ vvl_sources = [
   "layers/best_practices/bp_buffer.cpp",
   "layers/best_practices/bp_cmd_buffer.cpp",
   "layers/best_practices/bp_cmd_buffer_nv.cpp",
+  "layers/best_practices/bp_constants.h",
   "layers/best_practices/bp_copy_blit_resolve.cpp",
   "layers/best_practices/bp_descriptor.cpp",
   "layers/best_practices/bp_device_memory.cpp",
@@ -87,7 +88,9 @@ vvl_sources = [
   "layers/best_practices/bp_pipeline.cpp",
   "layers/best_practices/bp_ray_tracing.cpp",
   "layers/best_practices/bp_render_pass.cpp",
+  "layers/best_practices/bp_state_tracker.cpp",
   "layers/best_practices/bp_state.h",
+  "layers/best_practices/bp_state.cpp",
   "layers/best_practices/bp_synchronization.cpp",
   "layers/best_practices/bp_video.cpp",
   "layers/best_practices/bp_wsi.cpp",
