+++ layers/best_practices/bp_instance_device.cpp
@@ -21,6 +21,7 @@
 #include "generated/dispatch_functions.h"
 #include "best_practices/bp_state.h"
 #include "state_tracker/queue_state.h"
+#include "state_tracker/device_state.h"
 
 bool bp_state::Instance::ValidateDeprecatedExtensions(const Location& loc, vvl::Extension extension, APIVersion version) const {
     bool skip = false;
@@ -370,10 +371,10 @@ bool BestPractices::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindI
         const Location bind_info_loc = error_obj.location.dot(Field::pBindInfo, bind_idx);
         const VkBindSparseInfo& bind_info = pBindInfo[bind_idx];
         // Store sparse binding image_state and after binding is complete make sure that any requiring metadata have it bound
-        vvl::unordered_set<const vvl::Image*> sparse_images;
+        vvl::unordered_set<const bp_state::ImageSubState*> sparse_images;
         // Track images getting metadata bound by this call in a set, it'll be recorded into the image_state
         // in RecordQueueBindSparse.
-        vvl::unordered_set<const vvl::Image*> sparse_images_with_metadata;
+        vvl::unordered_set<const bp_state::ImageSubState*> sparse_images_with_metadata;
         // If we're binding sparse image memory make sure reqs were queried and note if metadata is required and bound
         for (uint32_t i = 0; i < bind_info.imageBindCount; ++i) {
             const auto& image_bind = bind_info.pImageBinds[i];
@@ -381,9 +382,10 @@ bool BestPractices::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindI
             if (!image_state) {
                 continue;  // Param/Object validation should report image_bind.image handles being invalid, so just skip here.
             }
-            sparse_images.insert(image_state.get());
+            const auto& image_sub_state = bp_state::SubState(*image_state);
+            sparse_images.insert(&image_sub_state);
             if (image_state->sparse_residency) {
-                if (!image_state->get_sparse_reqs_called || image_state->sparse_requirements.empty()) {
+                if (!image_sub_state.get_sparse_reqs_called || image_state->sparse_requirements.empty()) {
                     // For now just warning if sparse image binding occurs without calling to get reqs first
                     skip |= LogWarning("BestPractices-vkQueueBindSparse-image-requirements2", image_state->Handle(),
                                        bind_info_loc.dot(Field::pImageBinds, i),
@@ -392,7 +394,7 @@ bool BestPractices::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindI
                                        FormatHandle(image_state->Handle()).c_str());
                 }
             }
-            if (!image_state->memory_requirements_checked[0]) {
+            if (!image_sub_state.memory_requirements_checked[0]) {
                 // For now just warning if sparse image binding occurs without calling to get reqs first
                 skip |= LogWarning("BestPractices-vkQueueBindSparse-image-requirements", image_state->Handle(),
                                    bind_info_loc.dot(Field::pImageBinds, i),
@@ -407,9 +409,10 @@ bool BestPractices::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindI
             if (!image_state) {
                 continue;  // Param/Object validation should report image_bind.image handles being invalid, so just skip here.
             }
-            sparse_images.insert(image_state.get());
+            const auto& image_sub_state = bp_state::SubState(*image_state);
+            sparse_images.insert(&image_sub_state);
             if (image_state->sparse_residency) {
-                if (!image_state->get_sparse_reqs_called || image_state->sparse_requirements.empty()) {
+                if (!image_sub_state.get_sparse_reqs_called || image_state->sparse_requirements.empty()) {
                     // For now just warning if sparse image binding occurs without calling to get reqs first
                     skip |= LogWarning("BestPractices-vkQueueBindSparse-image-opaque-requirements2", image_state->Handle(),
                                        bind_info_loc.dot(Field::pImageOpaqueBinds, i),
@@ -418,7 +421,7 @@ bool BestPractices::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindI
                                        FormatHandle(image_state->Handle()).c_str());
                 }
             }
-            if (!image_state->memory_requirements_checked[0]) {
+            if (!image_sub_state.memory_requirements_checked[0]) {
                 // For now just warning if sparse image binding occurs without calling to get reqs first
                 skip |= LogWarning("BestPractices-vkQueueBindSparse-image-opaque-requirements", image_state->Handle(),
                                    bind_info_loc.dot(Field::pImageOpaqueBinds, i),
@@ -428,7 +431,7 @@ bool BestPractices::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindI
             }
             for (uint32_t j = 0; j < image_opaque_bind.bindCount; ++j) {
                 if (image_opaque_bind.pBinds[j].flags & VK_SPARSE_MEMORY_BIND_METADATA_BIT) {
-                    sparse_images_with_metadata.insert(image_state.get());
+                    sparse_images_with_metadata.insert(&image_sub_state);
                 }
             }
         }
@@ -436,11 +439,11 @@ bool BestPractices::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindI
             if (sparse_image_state->sparse_metadata_required && !sparse_image_state->sparse_metadata_bound &&
                 sparse_images_with_metadata.find(sparse_image_state) == sparse_images_with_metadata.end()) {
                 // Warn if sparse image binding metadata required for image with sparse binding, but metadata not bound
-                skip |= LogWarning("BestPractices-vkQueueBindSparse-image-metadata-requirements", sparse_image_state->Handle(),
+                skip |= LogWarning("BestPractices-vkQueueBindSparse-image-metadata-requirements", sparse_image_state->base.Handle(),
                                    bind_info_loc,
                                    "Binding sparse memory to %s which requires a metadata aspect but no "
                                    "binding with VK_SPARSE_MEMORY_BIND_METADATA_BIT set was made.",
-                                   FormatHandle(sparse_image_state->Handle()).c_str());
+                                   FormatHandle(sparse_image_state->base.Handle()).c_str());
             }
         }
     }
@@ -473,9 +476,10 @@ void BestPractices::ManualPostCallRecordQueueBindSparse(VkQueue queue, uint32_t 
             if (!image_state) {
                 continue;  // Param/Object validation should report image_bind.image handles being invalid, so just skip here.
             }
+            auto& image_sub_state = bp_state::SubState(*image_state);
             for (uint32_t j = 0; j < image_opaque_bind.bindCount; ++j) {
                 if (image_opaque_bind.pBinds[j].flags & VK_SPARSE_MEMORY_BIND_METADATA_BIT) {
-                    image_state->sparse_metadata_bound = true;
+                    image_sub_state.sparse_metadata_bound = true;
                 }
             }
         }
