+++ layers/state_tracker/image_state.h
@@ -102,14 +102,9 @@ class Image : public Bindable, public SubStateManager<ImageSubState> {
     static constexpr int kMaxPlanes = 3;
     using MemoryReqs = std::array<VkMemoryRequirements, kMaxPlanes>;
     const MemoryReqs requirements;
-    std::array<bool, kMaxPlanes> memory_requirements_checked = {};
 
     const bool sparse_residency;
-    using SparseReqs = std::vector<VkSparseImageMemoryRequirements>;
-    const SparseReqs sparse_requirements;
-    const bool sparse_metadata_required;  // Track if sparse metadata aspect is required for this image
-    bool get_sparse_reqs_called;          // Track if GetImageSparseMemoryRequirements() has been called for this image
-    bool sparse_metadata_bound;           // Track if sparse metadata aspect is bound to this image
+    const std::vector<VkSparseImageMemoryRequirements> sparse_requirements;
 
     VkImageFormatProperties image_format_properties = {};
 #ifdef VK_USE_PLATFORM_METAL_EXT
