+++ layers/state_tracker/state_tracker.cpp
@@ -1535,72 +1535,6 @@ void DeviceState::PostCallRecordBindBufferMemory2KHR(VkDevice device, uint32_t b
     PostCallRecordBindBufferMemory2(device, bindInfoCount, pBindInfos, record_obj);
 }
 
-void DeviceState::RecordGetImageMemoryRequirementsState(VkImage image, const VkImageMemoryRequirementsInfo2 *pInfo) {
-    const VkImagePlaneMemoryRequirementsInfo *plane_info =
-        (pInfo == nullptr) ? nullptr : vku::FindStructInPNextChain<VkImagePlaneMemoryRequirementsInfo>(pInfo->pNext);
-    if (auto image_state = Get<Image>(image)) {
-        if (plane_info != nullptr) {
-            // Multi-plane image
-            if (plane_info->planeAspect == VK_IMAGE_ASPECT_PLANE_0_BIT) {
-                image_state->memory_requirements_checked[0] = true;
-            } else if (plane_info->planeAspect == VK_IMAGE_ASPECT_PLANE_1_BIT) {
-                image_state->memory_requirements_checked[1] = true;
-            } else if (plane_info->planeAspect == VK_IMAGE_ASPECT_PLANE_2_BIT) {
-                image_state->memory_requirements_checked[2] = true;
-            }
-        } else if (!image_state->disjoint) {
-            // Single Plane image
-            image_state->memory_requirements_checked[0] = true;
-        }
-    }
-}
-
-void DeviceState::PostCallRecordGetImageMemoryRequirements(VkDevice device, VkImage image,
-                                                           VkMemoryRequirements *pMemoryRequirements,
-                                                           const RecordObject &record_obj) {
-    RecordGetImageMemoryRequirementsState(image, nullptr);
-}
-
-void DeviceState::PostCallRecordGetImageMemoryRequirements2(VkDevice device, const VkImageMemoryRequirementsInfo2 *pInfo,
-                                                            VkMemoryRequirements2 *pMemoryRequirements,
-                                                            const RecordObject &record_obj) {
-    RecordGetImageMemoryRequirementsState(pInfo->image, pInfo);
-}
-
-void DeviceState::PostCallRecordGetImageMemoryRequirements2KHR(VkDevice device, const VkImageMemoryRequirementsInfo2 *pInfo,
-                                                               VkMemoryRequirements2 *pMemoryRequirements,
-                                                               const RecordObject &record_obj) {
-    RecordGetImageMemoryRequirementsState(pInfo->image, pInfo);
-}
-
-void DeviceState::PostCallRecordGetImageSparseMemoryRequirements(VkDevice device, VkImage image,
-                                                                 uint32_t *pSparseMemoryRequirementCount,
-                                                                 VkSparseImageMemoryRequirements *pSparseMemoryRequirements,
-                                                                 const RecordObject &record_obj) {
-    if (auto image_state = Get<Image>(image)) {
-        image_state->get_sparse_reqs_called = true;
-    }
-}
-
-void DeviceState::PostCallRecordGetImageSparseMemoryRequirements2(VkDevice device,
-                                                                  const VkImageSparseMemoryRequirementsInfo2 *pInfo,
-                                                                  uint32_t *pSparseMemoryRequirementCount,
-                                                                  VkSparseImageMemoryRequirements2 *pSparseMemoryRequirements,
-                                                                  const RecordObject &record_obj) {
-    if (auto image_state = Get<Image>(pInfo->image)) {
-        image_state->get_sparse_reqs_called = true;
-    }
-}
-
-void DeviceState::PostCallRecordGetImageSparseMemoryRequirements2KHR(VkDevice device,
-                                                                     const VkImageSparseMemoryRequirementsInfo2 *pInfo,
-                                                                     uint32_t *pSparseMemoryRequirementCount,
-                                                                     VkSparseImageMemoryRequirements2 *pSparseMemoryRequirements,
-                                                                     const RecordObject &record_obj) {
-    PostCallRecordGetImageSparseMemoryRequirements2(device, pInfo, pSparseMemoryRequirementCount, pSparseMemoryRequirements,
-                                                    record_obj);
-}
-
 void DeviceState::PreCallRecordDestroyShaderModule(VkDevice device, VkShaderModule shaderModule,
                                                    const VkAllocationCallbacks *pAllocator, const RecordObject &record_obj) {
     Destroy<ShaderModule>(shaderModule);
